gears/mriqc-non-static/requirements.txt
atomicwrites==1.4.1 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
attrs==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
beautifulsoup4==4.12.2 ; python_version >= "3.8" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.3.0 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
dotty-dict==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
flywheel-gear-toolkit==0.6.13 ; python_version >= "3.8" and python_version < "4.0"
flywheel-gears==0.3.0 ; python_version >= "3.8" and python_version < "4.0"
flywheel-sdk==17.5.0 ; python_version >= "3.8" and python_version < "4.0"
fw-utils==4.3.5 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==3.2.0 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.8" and python_version < "4.0"
pluggy==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.8" and python_version < "4.0"
pyfakefs==5.3.0 ; python_version >= "3.8" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.8" and python_version < "4.0"
pytest-subprocess==1.5.0 ; python_version >= "3.8" and python_version < "4.0"
pytest==6.2.5 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.8" and python_version < "4.0"
rfc3987==1.3.8 ; python_version >= "3.8" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.8" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.8" and python_version < "4.0"
tzlocal==5.2 ; python_version >= "3.8" and python_version < "4.0"
urllib3==2.0.6 ; python_version >= "3.8" and python_version < "4.0"
